Fix sdl2 initial paint
On a tiling window manager, initial paint of sdl window can be missed, leaving
the window in an unpainted state, until triggered through resize/input. It seems
a SDL_WINDOWEVENT_SIZE_CHANGED event is fired upon window spawn with tilers,
instead of SDL_WINDOWEVENT_RESIZED, causing the paint to be missed when a race
is lost.